+++ EssentialFeed/EssentialFeed/Feed API/FeedItemsMapper.swift
@@ -7,13 +7,6 @@
 
 import Foundation
 
-internal struct RemoteFeedItem: Decodable {
-	internal let id: UUID
-	internal let description: String?
-	internal let location: String?
-	internal let image: URL
-}
-
 final class FeedItemsMapper {
 	private struct Root: Decodable {
 		let items: [RemoteFeedItem]
